Consensus: Rename: consensus/utilmoneystr.cpp belongs in libconsensus because FormatMoney()
- My opinion: there's only satoshis (aka units) in libconsensus
