+++ src/wallet/wallet.cpp
@@ -13,6 +13,7 @@
 #include "consensus/consensus.h"
 #include "consensus/script.h"
 #include "consensus/transaction.h"
+#include "consensus/utilmoneystr.h"
 #include "consensus/validation.h"
 #include "key.h"
 #include "keystore.h"
@@ -24,7 +25,6 @@
 #include "txmempool.h"
 #include "util.h"
 #include "ui_interface.h"
-#include "utilmoneystr.h"
 
 #include <assert.h>
